# backend/docs/swagger.yaml
basePath: /
definitions:
...
  dto.CreateSubflagRequest:
    properties:
      name:
        type: string
      sortOrder:
        type: integer
    type: object
  dto.CreateTaskRequest:
    properties:
      description:
        type: string
      dueAt:
        type: string
      status:
        type: string
      title:
        type: string
    type: object
  dto.ErrorResponse:
    properties:
      error:
        type: string
      requestId:
        type: string
    type: object
...
  dto.ListSubflagsResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.SubflagResponse'
        type: array
      nextCursor:
        type: string
    type: object
  dto.ListTasksResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.TaskResponse'
        type: array
      nextCursor:
        type: string
    type: object
...
  dto.SubflagResponse:
    properties:
      createdAt:
        type: string
      flagId:
        type: string
      id:
        type: string
      name:
        type: string
      sortOrder:
        type: integer
      updatedAt:
        type: string
    type: object
  dto.TaskResponse:
    properties:
      createdAt:
        type: string
      description:
        type: string
      dueAt:
        type: string
      id:
        type: string
      sourceInboxItemId:
        type: string
      status:
        type: string
      title:
        type: string
      updatedAt:
        type: string
    type: object
...
  dto.UpdateSubflagRequest:
    properties:
      name:
        type: string
      sortOrder:
        type: integer
    type: object
  dto.UpdateTaskRequest:
    properties:
      description:
        type: string
      dueAt:
        type: string
      status:
        type: string
      title:
        type: string
    type: object
...
info:
  contact: {}
  description: API do MVP Inbota.
  title: Inbota API
  version: "0.1"
paths:
...
  /v1/flags/{id}/subflags:
    get:
      parameters:
      - description: Flag ID
        in: path
        name: id
        required: true
        type: string
      - description: Limite
        in: query
        name: limit
        type: integer
      - description: Cursor
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ListSubflagsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Listar subflags por flag
      tags:
      - Subflags
    post:
      consumes:
      - application/json
      parameters:
      - description: Flag ID
        in: path
        name: id
        required: true
        type: string
      - description: Subflag payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateSubflagRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.SubflagResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Criar subflag
      tags:
      - Subflags
...
  /v1/subflags/{id}:
    delete:
      parameters:
      - description: Subflag ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Remover subflag
      tags:
      - Subflags
    patch:
      consumes:
      - application/json
      parameters:
      - description: Subflag ID
        in: path
        name: id
        required: true
        type: string
      - description: Subflag payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateSubflagRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SubflagResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Atualizar subflag
      tags:
      - Subflags
  /v1/tasks:
    get:
      parameters:
      - description: Limite
        in: query
        name: limit
        type: integer
      - description: Cursor
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ListTasksResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Listar tarefas
      tags:
      - Tasks
    post:
      consumes:
      - application/json
      parameters:
      - description: Task payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.TaskResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Criar tarefa
      tags:
      - Tasks
  /v1/tasks/{id}:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      - description: Task payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TaskResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Atualizar tarefa
      tags:
      - Tasks
securityDefinitions:
  BearerAuth:
    description: 'Formato: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"